refactor out ccsds components into their own sub-directory; validate with tests; improve test suite to display totals
